Moving the class HistoricConversionQueryBuilder to right place
