+++ src/test/java/org/javamoney/moneta/internal/convert/ECBHistoricRateProviderTest.java
@@ -35,8 +35,8 @@ import javax.money.convert.CurrencyConversion;
 import javax.money.convert.ExchangeRateProvider;
 
 import org.javamoney.moneta.ExchangeRateType;
+import org.javamoney.moneta.HistoricConversionQueryBuilder;
 import org.javamoney.moneta.Money;
-import org.javamoney.moneta.convert.HistoricConversionQueryBuilder;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
